All changes needed at least to compile with the new C2 refactoring: - All missing fields in beacon that are now in its transport - Same for session

+++ server/generate/c2.go
@@ -50,7 +50,7 @@ func InitImplantTransports(pbConfig *clientpb.ImplantConfig, cfg *models.Implant
 		if err != nil {
 			return err
 		}
-		transport.Priority = order
+		transport.Priority = int32(order)
 		cfg.Transports = append(cfg.Transports, transport)
 	}
